movefiltering-adjacent tests to separate test file
